:tada: feat: improve function
